requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
peewee==3.17.7
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
urllib3==2.2.3
